feat: url에 따라 게시글과 동영상과 정보 표시 (#19)
