+++ src/api/video.js
@@ -1,6 +1,8 @@
 import { baseApi } from './base';
 
-export const getVideo = async (url) => {
-  const response = await baseApi.get(url);
+export const getVideo = async (videoId) => {
+  const response = await baseApi.get(
+    `videos?part=snippet%2Cstatistics&id=${videoId}&key=${process.env.REACT_APP_ACCESS_TOKEN}`
+  );
   return response.data;
 };
