+++ src/components/Detail/Sidebar.jsx
@@ -1,7 +1,6 @@
-import React, { useState, useEffect } from 'react';
+import React, { useState } from 'react';
 import { RecommendMovies, PostInfo, Comments } from 'components/DetailPost';
 import { StSidebar, StSidebarNav } from './styles';
-import { getPost } from 'api/post';
 
 const tabs = {
   recommend: '추천',
@@ -9,19 +8,11 @@ const tabs = {
   comments: '댓글'
 };
 
-export default function Sidebar({ postId }) {
-  const [postInfo, setPostInfo] = useState({});
+export default function Sidebar({ postInfo, postId }) {
   const [selectedTab, setSelectedTab] = useState(tabs.recommend);
   const handleClickBtn = (e) => setSelectedTab(e.target.name);
 
-  useEffect(() => {
-    const fetchData = async () => {
-      const res = await getPost(postId);
-      setPostInfo(res);
-    };
-
-    fetchData();
-  }, [postId]);
+  if (!postInfo) return;
 
   return (
     <StSidebar>
