+++ src/components/Detail/index.js
@@ -1,5 +1,6 @@
 import VideoInfo from './VideoInfo';
 import VideoPlayer from './VideoPlayer';
 import Sidebar from './Sidebar';
+import FetchLoader from './FetchLoader';
 
-export { VideoInfo, VideoPlayer, Sidebar };
+export { VideoInfo, VideoPlayer, Sidebar, FetchLoader };
